+++ src/app/components/card-list/card-list.scss
