+++ src/app/components/card-list/card/card.scss
