+++ src/app/components/card-list/card/device/device.scss
