+++ src/app/components/card-list/card/sensor/sensor.spec.ts
