+++ src/app/components/card-list/device-single/device-single.html
