+++ src/app/components/sidebar/footer/sidebar-footer.html
