+++ src/app/components/sidebar/footer/sidebar-footer.spec.ts
