+++ src/app/components/sidebar/header/sidebar-header.scss
