+++ src/app/components/sidebar/menu/menu.html
