+++ src/app/components/sidebar/menu/menu.scss
