+++ src/app/components/sidebar/sidebar.scss
