+++ src/app/entity/directives/card-highlight.spec.ts
