+++ src/app/entity/directives/card-highlight.ts
