+++ src/app/entity/directives/lamp-highlight.spec.ts
