+++ src/app/entity/directives/lamp-highlight.ts
