+++ src/app/entity/pipes/sensor-value-pipe.spec.ts
