+++ src/app/entity/pipes/sensor-value-pipe.ts
