+++ src/app/pages/dashboard/dashboard.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { MatTabsModule } from '@angular/material/tabs';
 import type { TabData } from '~/api/api.types';
 import { api } from '~/api/api.service';
-import { CardListComponent } from './card-list/card-list';
+import { CardListComponent } from 'ui/card-list/card-list';
 
 @Component({
   selector: 'app-dashboard',
